# openapi.orca.yaml
openapi: 3.0.3
info:
  title: ORCA AgentMesh API
  description: |
    ORCA AgentMesh API for managing agents, workflows, trust scoring, and observability.
    
    Features:
    - UADSI trust scoring and risk metrics
    - MCP agent discovery and alignment
    - Adapter webhooks (Zapier, n8n)
    - Self-healing diagnostics
    - Executive reporting
  version: 1.0.0
  contact:
    name: ORCA AgentMesh Team
    email: support@orca.agentmesh.dev
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.orca.agentmesh.dev
    description: Production server
...
paths:
...
  /workflows/{id}:
    get:
      summary: Get workflow by ID
      tags: [workflows]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workflow details
        '404':
          description: Workflow not found
  /workflows/{id}/events:
    get:
      summary: Get workflow events
      tags: [workflows]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Workflow events
...
  /adapters/zapier/webhook:
    post:
      summary: Zapier webhook endpoint
      tags: [adapters]
      security:
        - hmacSignature: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZapierWebhook'
      responses:
        '200':
          description: Webhook processed
        '401':
          description: Invalid signature
  /adapters/n8n/webhook:
    post:
      summary: n8n webhook endpoint
      tags: [adapters]
      security:
        - hmacSignature: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/N8NWebhook'
      responses:
        '200':
          description: Webhook processed
        '401':
          description: Invalid signature
components:
  schemas:
...
    ZapierWebhook:
      type: object
      properties:
        zap_id:
          type: string
        zap_name:
          type: string
        event_type:
          type: string
        data:
          type: object
        timestamp:
          type: string
    N8NWebhook:
      type: object
      properties:
        workflow_id:
          type: string
        workflow_name:
          type: string
        execution_id:
          type: string
        event:
          type: string
        data:
          type: object
        timestamp:
          type: string
  securitySchemes:
    hmacSignature:
      type: apiKey
      in: header
      name: x-signature
      description: HMAC SHA-256 signature of request body
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT